Solved issue with UpgradeIT being dependent on config on host
