+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/ClientIT.java
@@ -16,26 +16,26 @@
 
 package io.dockstore.client.cli;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.concurrent.TimeoutException;
-
-import org.apache.commons.dbutils.handlers.ScalarHandler;
-import org.apache.commons.io.FileUtils;
-import org.junit.*;
-import org.junit.contrib.java.lang.system.ExpectedSystemExit;
-
-import com.google.common.io.Files;
-
 import io.dockstore.common.CommonTestUtilities.TestingPostgres;
+import io.dockstore.common.TestUtility;
 import io.dockstore.webservice.DockstoreWebserviceApplication;
 import io.dockstore.webservice.DockstoreWebserviceConfiguration;
 import io.dockstore.webservice.core.Registry;
 import io.dropwizard.testing.ResourceHelpers;
 import io.dropwizard.testing.junit.DropwizardAppRule;
 import io.swagger.client.ApiException;
+import org.apache.commons.dbutils.handlers.ScalarHandler;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Ignore;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.contrib.java.lang.system.ExpectedSystemExit;
 
-import static io.dockstore.common.CommonTestUtilities.DUMMY_TOKEN_1;
+import java.io.IOException;
+import java.util.concurrent.TimeoutException;
+
 import static io.dockstore.common.CommonTestUtilities.clearState;
 import static io.dockstore.common.CommonTestUtilities.getTestingPostgres;
 
@@ -57,45 +57,32 @@ public class ClientIT {
         clearState();
     }
 
-    public static String getConfigFileLocation(boolean correctUser) throws IOException {
-        return getConfigFileLocation(correctUser, true);
-    }
-
-    public static String getConfigFileLocation(boolean correctUser, boolean validPort) throws IOException {
-        File tempDir = Files.createTempDir();
-        final File tempFile = File.createTempFile("config", "config", tempDir);
-        FileUtils.write(tempFile, "token: " + (correctUser ? DUMMY_TOKEN_1 : "foobar") + "\n");
-        FileUtils.write(tempFile, "server-url: http://localhost:" + (validPort ? "8000" : "9001") + "\n", true);
-
-        return tempFile.getAbsolutePath();
-    }
-
     @Test
     public void testListEntries() throws IOException, TimeoutException, ApiException {
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "list" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "list" });
     }
 
     @Test
     public void testDebugModeListEntries() throws IOException, TimeoutException, ApiException {
-        Client.main(new String[] { "--debug", "--config", getConfigFileLocation(true), "tool", "list" });
+        Client.main(new String[] { "--debug", "--config", TestUtility.getConfigFileLocation(true), "tool", "list" });
     }
 
     @Test
     public void testListEntriesWithoutCreds() throws IOException, TimeoutException, ApiException {
         systemExit.expectSystemExitWithStatus(Client.API_ERROR);
-        Client.main(new String[] { "--config", getConfigFileLocation(false), "tool", "list" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(false), "tool", "list" });
     }
 
     @Test
     public void testListEntriesOnWrongPort() throws IOException, TimeoutException, ApiException {
         systemExit.expectSystemExitWithStatus(Client.CONNECTION_ERROR);
-        Client.main(new String[] { "--config", getConfigFileLocation(true, false), "tool", "list" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true, false), "tool", "list" });
     }
 
     // Won't work as entry must be valid
     @Ignore
     public void quickRegisterValidEntry() throws IOException {
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6" });
 
         // verify DB
         final TestingPostgres testingPostgres = getTestingPostgres();
@@ -105,9 +92,9 @@ public class ClientIT {
 
     @Ignore
     public void quickRegisterDuplicateEntry() throws IOException {
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6", "view1" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6", "view2" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6", "view1" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test6", "view2" });
 
         // verify DB
         final TestingPostgres testingPostgres = getTestingPostgres();
@@ -118,13 +105,13 @@ public class ClientIT {
     @Test
     public void quickRegisterInValidEntry() throws IOException {
         systemExit.expectSystemExitWithStatus(Client.CLIENT_ERROR);
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test1" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "publish", "quay.io/test_org/test1" });
     }
 
     @Test
     public void quickRegisterUnknownEntry() throws IOException {
         systemExit.expectSystemExitWithStatus(Client.CLIENT_ERROR);
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "publish", "quay.io/funky_container_that_does_not_exist" });
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "publish", "quay.io/funky_container_that_does_not_exist" });
     }
 
     /* When you manually publish on the dockstore CLI, it will now refresh the container after it is added.
@@ -133,19 +120,19 @@ public class ClientIT {
      */
     @Ignore("Since dockstore now checks for associated tags for Quay container, manual publishing of nonexistant images won't work")
     public void manualRegisterABunchOfValidEntries() throws IOException {
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master", "--toolname", "test1" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master", "--toolname", "test2" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.DOCKER_HUB.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.DOCKER_HUB.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.DOCKER_HUB.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.DOCKER_HUB.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master", "--toolname", "test1" });
 
@@ -159,13 +146,13 @@ public class ClientIT {
     @Test
     public void manualRegisterADuplicate() throws IOException {
         systemExit.expectSystemExitWithStatus(Client.API_ERROR);
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master", "--toolname", "test1" });
-        Client.main(new String[] { "--config", getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "manual_publish", "--registry", Registry.QUAY_IO.toString(),
                 "--namespace", "pypi", "--name", "bd2k-python-lib", "--git-url", "git@github.com:funky-user/test2.git", "--git-reference",
                 "refs/head/master", "--toolname", "test1" });
     }
