+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/MockedIT.java
@@ -107,19 +107,19 @@ public class MockedIT {
 
     @Ignore
     public void runLaunchOneJson() throws IOException, ApiException {
-        Client.main(new String[] { "--config", ClientIT.getConfigFileLocation(true), "tool", "launch", "--entry",
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "launch", "--entry",
             "quay.io/collaboratory/dockstore-tool-linux-sort", "--json", ResourceHelpers.resourceFilePath("testOneRun.json"), "--debug" });
     }
 
     @Ignore
     public void runLaunchNJson() throws IOException {
-        Client.main(new String[] { "--config", ClientIT.getConfigFileLocation(true), "tool", "launch", "--entry",
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "launch", "--entry",
                 "quay.io/collaboratory/dockstore-tool-linux-sort", "--json", ResourceHelpers.resourceFilePath("testMultipleRun.json") });
     }
 
     @Ignore
     public void runLaunchTSV() throws IOException {
-        Client.main(new String[] { "--config", ClientIT.getConfigFileLocation(true), "tool", "launch", "--entry",
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "launch", "--entry",
                 "quay.io/collaboratory/dockstore-tool-linux-sort", "--tsv", ResourceHelpers.resourceFilePath("testMultipleRun.tsv") });
     }
 
@@ -130,7 +130,7 @@ public class MockedIT {
      */
     @Ignore
     public void runLaunchOneLocalArrayedJson() throws IOException, ApiException {
-        Client.main(new String[] { "--config", ClientIT.getConfigFileLocation(true), "tool", "launch", "--entry",
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "launch", "--entry",
             "quay.io/collaboratory/arrays", "--json", ResourceHelpers.resourceFilePath("testArrayLocalInputLocalOutput.json") });
 
         Assert.assertTrue(new File("/tmp/example.bedGraph").exists());
@@ -143,8 +143,8 @@ public class MockedIT {
      */
     @Ignore
     public void runLaunchOneHTTPArrayedJson() throws IOException, ApiException {
-        System.out.println(ClientIT.getConfigFileLocation(true));
-        Client.main(new String[] { "--config", ClientIT.getConfigFileLocation(true), "tool", "launch", "--entry",
+        System.out.println(TestUtility.getConfigFileLocation(true));
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "launch", "--entry",
             "quay.io/collaboratory/arrays", "--json", ResourceHelpers.resourceFilePath("testArrayHttpInputLocalOutput.json") });
 
         Assert.assertTrue(new File("/tmp/wc1.out").exists());
@@ -159,7 +159,7 @@ public class MockedIT {
          */
     @Test
     public void testRun() throws IOException, ApiException {
-        System.out.println(ClientIT.getConfigFileLocation(true));
-        Client.main(new String[] { "--config", ClientIT.getConfigFileLocation(true), "tool", "list" });
+        System.out.println(TestUtility.getConfigFileLocation(true));
+        Client.main(new String[] { "--config", TestUtility.getConfigFileLocation(true), "tool", "list" });
     }
 }
